+++ docs/CHANGELOG.md
@@ -3,8 +3,8 @@
 
 ## 2.3.8 -> 2.3.9:
 * Main improvements:
-  * New documentation
-  * Move from jshint to eslint
+  * Documentation updated (move to wepsim.github.io/info/)
+  * Update from jshint to eslint
 
 ## 2.3.7 -> 2.3.8:
 * Main improvements: